requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# API and web frameworks
flask==3.0.0
fastapi==0.104.1

# Machine Learning (uncomment if needed)
# scikit-learn==1.3.2
# tensorflow==2.15.0
# torch==2.1.2

# Development tools
pytest==7.4.3
black==23.12.1
flake8==6.1.0

# Add your specific dependencies here
